refactor: create fan_controller.py with mixin-based FanController, keep logic.py as re-export
